#225 chore: 엔티티 수정
