+++ src/main/java/com/codiary/backend/domain/member/entity/Follow.java
@@ -3,7 +3,6 @@ package com.codiary.backend.domain.member.entity;
 import com.codiary.backend.global.common.BaseEntity;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
@@ -12,8 +11,7 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class Follow extends BaseEntity{
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "follow_id", nullable = false, columnDefinition = "bigint")
     private Long followId;
 
@@ -27,15 +25,4 @@ public class Follow extends BaseEntity{
 
     @Column(name="follow_status", columnDefinition = "tinyint")
     private Boolean followStatus;
-
-    @Builder
-    public Follow(Member fromMember, Member toMember, Boolean followStatus) {
-        this.fromMember = fromMember;
-        this.toMember = toMember;
-        this.followStatus = followStatus;
-    }
-
-    public void update(Boolean followStatus) {
-        this.followStatus = followStatus;
-    }
 }
