+++ src/main/java/com/codiary/backend/domain/member/entity/Uuid.java
@@ -4,15 +4,18 @@ import jakarta.persistence.*;
 import lombok.*;
 
 @Entity
-@Builder
 @Getter
-@AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class Uuid {
 
-  @Id  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
   private Long id;
 
   @Column(name = "uuid", columnDefinition = "varchar(256)")
   private String uuid;
+
+  @Builder
+  public Uuid(String uuid) {
+    this.uuid = uuid;
+  }
 }
