+++ src/main/java/com/codiary/backend/domain/post/entity/PostFile.java
@@ -4,10 +4,8 @@ import jakarta.persistence.*;
 import lombok.*;
 
 @Entity
-@Builder
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
 public class PostFile {
 
   @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -21,7 +19,7 @@ public class PostFile {
   @JoinColumn(name = "post_id")
   private Post post;
 
-  @Column(name = "file_url", nullable = false, columnDefinition = "varchar(500)")
+  @Column(nullable = false, columnDefinition = "varchar(500)")
   private String fileUrl;
 
 }
