+++ src/main/java/com/codiary/backend/domain/team/entity/TeamBannerImage.java
@@ -4,10 +4,8 @@ import jakarta.persistence.*;
 import lombok.*;
 
 @Entity
-@Builder
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
 public class TeamBannerImage {
 
     @Id
