+++ src/main/java/com/codiary/backend/domain/team/entity/TeamProfileImage.java
@@ -4,10 +4,8 @@ import jakarta.persistence.*;
 import lombok.*;
 
 @Entity
-@Builder
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
 public class TeamProfileImage {
 
     @Id
